[-] Refactor
